Split the explanation in common/override/multiple-urls paragraphs
I re-instated the original text and made clear that that is for *your own*
urls. This is the common case.

Second, you can override for instance django's URL.

Third: the part about args/kwargs, which was the original reason for the
original pull request.

I added a note about the resolution order.